modified layout to include scrollbars on sidebar and content i.e fixed sider
