+++ elakcustomerportal/src/index.css
@@ -24,3 +24,21 @@ code {
   border-right: 1px solid maroon;
   border-radius: 0;
 }
+
+.sider-dark {
+  background-color: #001529;
+}
+
+.sider-light {
+  background-color: #f5f5f5;
+}
+
+.sider-light .ant-menu {
+  white-space: nowrap;
+  overflow: auto;
+}
+
+.sider-dark .ant-menu {
+  white-space: nowrap;
+  overflow: auto;
+}
